\name{TestingR-package}
\alias{TestingR-package}
\alias{TestingR}
\docType{package}
\title{
\packageTitle{TestingR}
}
\description{
\packageDescription{TestingR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestingR}
\packageIndices{TestingR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TestingR}

Maintainer: \packageMaintainer{TestingR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
